pyproject.toml
[tool.poetry]
name = "smart_security_camera"
version = "0.1.0"
description = ""
authors = ["Adilson Pavan <adilson.pavan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.15.0"
paho-mqtt = "^1.5.1"
opencv-contrib-python = "^4.5.1"
python-telegram-bot = "^13.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
